feat: TossPaymentPayload 클래스의 amount 필드를 Long에서 Integer로 변경

+++ src/main/java/life/mosu/mosuserver/infra/payment/dto/TossPaymentPayload.java
@@ -1,7 +1,9 @@
 package life.mosu.mosuserver.infra.payment.dto;
 
 public record TossPaymentPayload(
-    String paymentKey,
-    String orderId,
-    Long amount
-){}
+        String paymentKey,
+        String orderId,
+        Integer amount
+) {
+
+}
